+++ lib/mangoapps/modules/tasks/get_task_details.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module MangoApps
+  class Client
+    module Tasks
+      module GetTaskDetails
+        def get_task_details(task_id, params = {})
+          get("tasks/#{task_id}.json", params: params)
+        end
+      end
+    end
+  end
+end