+++ lib/mangoapps/modules/tasks/get_tasks.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module MangoApps
+  class Client
+    module Tasks
+      module GetTasks
+        def get_tasks(params = {})
+          get("tasks/new_index.json", params: params)
+        end
+      end
+    end
+  end
+end